Refactor: 채팅방 정보 요청 로직 이동
