+++ components/chat/header.tsx
@@ -1,21 +1,21 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { HiArrowLongLeft } from 'react-icons/hi2';
 import { IoMdMenu } from 'react-icons/io';
-import { ChatUser } from '@/@types/types';
+import { Chat, ChatUser } from '@/@types/types';
+import chatAPI from '@/apis/chatAPI';
 import styles from './Chat.module.scss';
 import Jwtinterceptors from '../../apis/Jwtinterceptors';
 
 interface Props {
   chatId: string;
-  name: string;
-  users: ChatUser[];
 }
 
-export default function ChatroomHeader({ chatId, name, users }: Props) {
+export default function ChatroomHeader({ chatId }: Props) {
   const router = useRouter();
 
   const { instance } = Jwtinterceptors();
+  const [chatData, setChatData] = useState<Chat | null>();
 
   const accessToken: string = localStorage.getItem('accessToken');
 
@@ -52,45 +52,67 @@ export default function ChatroomHeader({ chatId, name, users }: Props) {
     }
   };
 
+  useEffect(() => {
+    const fetchChatData = async () => {
+      try {
+        if (chatId && typeof chatId === 'string') {
+          const response = await chatAPI.getChatInfo(chatId);
+          const chatInfo: Chat = response.data.chat;
+          setChatData(chatInfo);
+        }
+      } catch (error) {
+        console.error('Error fetching chat data:', error);
+      }
+    };
+
+    if (chatId) {
+      fetchChatData();
+    }
+  }, [chatId]);
+
   return (
     <div className={styles.header}>
       <div className={styles.left}>
         <HiArrowLongLeft onClick={handleBackBtnClick} />
       </div>
-      <h3 className={styles.chatTitle}>{name}</h3>
-      <div className={styles.right} onClick={toggleMenu}>
-        <IoMdMenu />
-        {/* Dropdown 메뉴 */}
-        {isMenuOpen && (
-          <div className={styles.dropdownMenu}>
-            <ul>
-              <div>
-                <p>현재 채팅인원 수</p>
-                <h6>{users.length}명</h6>
+      {chatData && (
+        <>
+          <h3 className={styles.chatTitle}>{chatData.name}</h3>
+          <div className={styles.right} onClick={toggleMenu}>
+            <IoMdMenu />
+            {/* Dropdown 메뉴 */}
+            {isMenuOpen && (
+              <div className={styles.dropdownMenu}>
+                <ul>
+                  <div>
+                    <p>현재 채팅인원 수</p>
+                    <h6>{chatData.users.length}명</h6>
+                  </div>
+                  <div className={styles.userDiv}>
+                    {chatData.users.map(user => (
+                      <li key={user.id}>
+                        <img
+                          src={user.picture}
+                          className={styles.profileImage}
+                          alt="User"
+                        />
+                        <span className={styles.username}>{user.username}</span>
+                      </li>
+                    ))}
+                  </div>
+                </ul>
+                <button
+                  type="submit"
+                  className={styles.exitButton}
+                  onClick={handleOutBtnClick}
+                >
+                  채팅방 나가기
+                </button>
               </div>
-              <div className={styles.userDiv}>
-                {users.map(user => (
-                  <li key={user.id}>
-                    <img
-                      src={user.picture}
-                      className={styles.profileImage}
-                      alt="User"
-                    />
-                    <span className={styles.username}>{user.username}</span>
-                  </li>
-                ))}
-              </div>
-            </ul>
-            <button
-              type="submit"
-              className={styles.exitButton}
-              onClick={handleOutBtnClick}
-            >
-              채팅방 나가기
-            </button>
+            )}
           </div>
-        )}
-      </div>
+        </>
+      )}
     </div>
   );
 }
